src/main/java/com/example/customer_api/services/CustomerUserDetailsService.java
package com.example.customer_api.services;

import com.example.customer_api.models.jpa.Customer;
import com.example.customer_api.models.service.CustomerDetail;
import com.example.customer_api.repo.CustomerRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
public class CustomerUserDetailsService implements UserDetailsService {

    @Autowired
    private CustomerRepository customerRepository;

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        Customer customer = customerRepository.findByUsername(username);
        if(customer == null) {
            throw new UsernameNotFoundException("User " + username + " not found.");
        }
        return new CustomerDetail(customer);
    }
}
